+++ CHANGELOG.md
@@ -7,6 +7,8 @@ For instructions on upgrading to newer versions, visit
 
 ### Resolved Issues
 
+* \#3243 Fixed `find_or_creeate_by` with custom persistence options.
+
 * \#3231 Allow evolution of proxy documents to work in criteria.
 
 * \#3247 Bump dependency on tzinfo to 0.3.29.